Use TypeScript and official DevTools Protocol typings.
