+++ .babelrc
@@ -1,3 +1,4 @@
 {
-  "plugins": [["@babel/plugin-transform-modules-commonjs"]]
+  "plugins": [["@babel/plugin-transform-modules-commonjs"]],
+  "presets": ["@babel/preset-typescript"]
 }
