+++ package.json
@@ -27,9 +27,11 @@
   "devDependencies": {
     "@babel/core": "^7.14.6",
     "@babel/plugin-transform-modules-commonjs": "^7.14.5",
+    "@babel/preset-typescript": "^7.15.0",
     "@pixi/jsdoc-template": "^2.6.0",
     "babel-jest": "^27.0.6",
     "css-loader": "^6.2.0",
+    "devtools-protocol": "^0.0.924232",
     "eslint": "^7.30.0",
     "eslint-config-google": "^0.14.0",
     "file-loader": "^6.2.0",
@@ -43,7 +45,9 @@
     "prettier": "^2.3.2",
     "puppeteer": "^9.1.1",
     "rimraf": "^3.0.2",
+    "source-map-loader": "^3.0.0",
     "style-loader": "^3.2.1",
+    "ts-loader": "^9.2.6",
     "webpack": "^5.44.0",
     "webpack-cli": "^4.7.2",
     "yazl": "^2.5.1"
@@ -53,6 +57,7 @@
     "build:rimraf": "rimraf build",
     "build:webpack": "webpack --mode production --config src/webpack.config.js",
     "build": "npm run build:rimraf && npm run build:webpack && npm run build:chrome-extension",
+    "dev": "webpack --mode development --config src/webpack.config.js && npm run build:chrome-extension",
     "postinstall": "husky install",
     "postpublish": "pinst --enable",
     "prepublishOnly": "pinst --disable",
