+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "target": "es2015",
+    "moduleResolution": "node"
+  },
+  "include": ["./src/**/*"]
+}
